pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "metapac"
version = "0.1.0"
description = "Meta-learning based Predictive Adaptive Compression for transformer models"
readme = "README.md"
authors = [
    {name = "Tamás Panyi", email = "panyi.tamas@nik.uni-obuda.hu"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = [
    "machine-learning",
    "deep-learning",
    "model-compression",
    "pruning",
    "quantization",
    "meta-learning",
    "transformers",
    "pytorch"
]

dependencies = [
    "numpy>=1.26",
    "pandas>=2.2",
    "scikit-learn>=1.4",
    "matplotlib>=3.8",
    "tqdm>=4.66",
    "torchmetrics>=1.4",
    "seaborn>=0.13",
    "datasets>=2.20",
    "accelerate>=0.33",
    "timm>=1.0.9",
    "rich",
    "torch>=2.0.1",
    "transformers>=4.44",
    "evaluate>=0.4",
    "tensorboard>=2.17",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.5",
    "isort>=5.12",
]

[project.urls]
Homepage = "https://github.com/alenoi/MetaPAC"
Repository = "https://github.com/alenoi/MetaPAC"
Issues = "https://github.com/alenoi/MetaPAC/issues"

[project.scripts]
metapac = "metapac.__main__:main"

[tool.setuptools]
packages = ["metapac", "metapac.src"]

[tool.setuptools.package-data]
metapac = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=metapac --cov-report=html --cov-report=term"
